Updated keras notebook
