+++ use-cases/nlp/classification/keras/code/requirements.txt
@@ -1,4 +1,3 @@
-neptune
 neptune-tensorflow-keras
 numpy
 pydot
