+++ use-cases/nlp/classification/keras/code/run_examples.sh
@@ -1,0 +1,7 @@
+set -e
+
+echo "Installing requirements..."
+pip install -U -r requirements.txt
+
+echo "Running keras_script.py..."
+python keras_script.py
